+++ src/mqtt/mqtt_usual.h
@@ -85,7 +85,7 @@ void onMqttDisconnect(AsyncMqttClientDisconnectReason reason)
     // Добавьте дополнительные случаи для анализа
 
     syslog_ng("mqtt: Disconnected. Reason: " + String(static_cast<int>(reason)));
-    syslog_ng("mqtt: WiFi isConnected: " + String(static_cast<int>(WiFi.isConnected())));
+    syslog_ng("mqtt: WiFi isConnected: " + String(WiFi.isConnected()));
     if (WiFi.isConnected())
     {
         if (not mqttClient.connected())
